utils/autoranker/requirements.txt
attrdict==2.0.0
beautifulsoup4==4.6.3
certifi==2018.8.13
chardet==3.0.4
cytoolz==0.9.0.1
eth-abi==1.1.1
eth-account==0.3.0
eth-hash==0.1.4
eth-keyfile==0.5.1
eth-keys==0.2.0b3
eth-rlp==0.1.2
eth-typing==1.1.0
eth-utils==1.2.0
hexbytes==0.1.0
idna==2.7
lru-dict==1.1.6
parsimonious==0.8.0
pycryptodome==3.6.6
requests==2.19.1
rlp==0.6.0
six==1.11.0
toolz==0.9.0
urllib3==1.23
web3>=4.5.0
websockets==5.0.1
py-geth>=1.4.0
py-solc>=0.4.0
pytest>=2.7.2
pysha3>=1.0.2
